requirements.txt
# Core ML and data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
scipy>=1.7.0
openpyxl>=3.1.0  # For Excel files
python-docx>=0.8.11  # For Word documents

# Deep Learning
torch>=1.9.0
transformers>=4.11.0

# Ensemble Methods
xgboost>=1.4.2
lightgbm>=3.2.1
catboost>=0.26.0

# Model Interpretability
shap>=0.39.0
lime>=0.2.0

# Microbiome specific
biom-format>=2.1.10
skbio>=0.5.6

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
httpx>=0.19.0  # For FastAPI testing

# Utilities
joblib>=1.0.1
tqdm>=4.62.3
python-dotenv>=0.19.0
loguru>=0.5.3

# Development
black>=21.9b0
flake8>=3.9.2
isort>=5.9.3
